fixed user name update
