+++ client/store/users.js
@@ -31,6 +31,7 @@ export function createUser(user) {
   return dispatch => {
     axios
       .post('/api/users', user)
+      .then(res => res.data)
       .then(user => {
         dispatch(addUser(user));
       })
@@ -38,16 +39,16 @@ export function createUser(user) {
   }
 }
 
-export function editUser(id, user) {
-  return dispatch => {
-    axios
-      .put(`/api/users/${id}`, user)
-      .then(user => {
-        dispatch(updateUser(user));
-      })
-      .catch(err => console.error("error updating user", err));
-  }
-}
+// export function editUser(id, user) {
+//   return dispatch => {
+//     axios
+//       .put(`/api/users/${id}`, user)
+//       .then(user => {
+//         dispatch(updateUser(user));
+//       })
+//       .catch(err => console.error("error updating user", err));
+//   }
+// }
 
 export function deleteAccount(id) {
   return dispatch => {
